packages/observability/pyproject.toml
[project]
name = "selva-observability"
version = "0.1.0"
description = "Shared logging and observability for Selva services"
requires-python = ">=3.12"
dependencies = [
  "structlog>=24.0",
]

[project.optional-dependencies]
sentry = ["sentry-sdk[fastapi]>=2.19.0"]
tracing = [
  "opentelemetry-api>=1.20",
  "opentelemetry-sdk>=1.20",
  "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]
all = [
  "selva-observability[sentry]",
  "selva-observability[tracing]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/selva_observability"]
